Watch and stop watching when not on friends screen
